+++ types/interface/ContextMenuOption.d.ts
@@ -1,4 +1,5 @@
 export default interface ContextMenuOption {
+    width?: number;
     hideMenuWhenScroll?: boolean;
     fixMenuWhenScroll?: boolean;
 }